# Remove performance regressions discovered in #1194