In #1194 I have analyzed and compared the results of all ML.NET end-to-end benchmarks for .NET Core 2.1 with native library for math and .NET Core 3.0 with managed CpuMath.

I found 3 quite huge regressions. After some extra profiling I found the reason:

**Most of the benchmarks use very short arrays, typically with just 4 or 14 floats. For such small arrays, the vectorization makes no sense.** This PR adds a very simple check: if the number of floats is small, don't run the vectorized code.  (/cc @AndyAyersMS this explains no perf boos in ML.NET scenarios after switch to new vectorized code)

|                             Type |              Method |     Toolchain |         Mean |      StdDev |
|--------------------------------- |-------------------- |-------------- |-------------:|------------:|
| KMeansAndLogisticRegressionBench |    TrainKMeansAndLR | netcoreapp2.1 |   566.601 ms | 168.8636 ms |
| KMeansAndLogisticRegressionBench |    TrainKMeansAndLR | netcoreapp3.0 |   568.968 ms |  92.8586 ms |
|            PredictionEngineBench | MakeIrisPredictions | netcoreapp2.1 |     3.896 ms |   0.0542 ms |
|            PredictionEngineBench | MakeIrisPredictions | netcoreapp3.0 |     3.820 ms |   0.0106 ms |


In 2.1 profiles we have calls to native dll:

![image](https://user-images.githubusercontent.com/6011991/47570276-11821800-d936-11e8-980d-458d7ea66da0.png)

Before my fix the code was calling the new managed CpuMath

![image](https://user-images.githubusercontent.com/6011991/47570345-3bd3d580-d936-11e8-81c5-e14312960a89.png)

Now there is no call to CpuMath if the vectors are small (and in most of our scenarios, they are). Moreover, the method gets inlined and we have no `pinvoke` cost compared to 2.1 and native dependency.

![image](https://user-images.githubusercontent.com/6011991/47570389-57d77700-d936-11e8-9c46-4fa2284734ad.png)

@danmosemsft I will need to take a look at the way ML.NET devides work into smaller tasks and change something on the upper level to take advantage of the vectorization.