@tannergooding 2018-10-29
I would also like to know if the small array case we have for the benchmarks is representative of real-world scenarios. I would not expect that most ML scenarios deal with small amounts of data as inputs to these methods, and we may be adding additional unnecessary checks that will negatively impact the real-world scenarios (even if minor). If small array inputs are fairly representative, then there are likely some other changes we could make to the vectorized code that would otherwise improve this more.