@tannergooding 2018-11-14
@adamsitnik, I am still waiting for a response on https://github.com/dotnet/machinelearning/pull/1394#issuecomment-433970212