@adamsitnik 2018-11-15
> I would also like to know if the small array case we have for the benchmarks is representative of real-world scenarios.

@tannergooding I am currently studying and profiling ML.NET to find the answer. I started with reading the input files where we have some bottlenecks. Once I got the parsing done right I am going to investigate and optimize the way we split data into smaller jobs to take advantage of CpuMath.