@shauheen 2018-11-16
@adamsitnik do you think it would be possible to close this PR and continue discussion in an issue until you have your investigation completed? then you can update the branch and submit a new PR?